fix include mysql.h Path. fix segmentation faults. By default directly return results.
The mysql header files are usually present in include folder and prefixing mysql doesn't work.
Date fields  can be null, and it was failing with 'Bus error', as null cannot be converted to String.
with mysql_use_result a new query cannot be executed until all the results from the previous query has been extracted. so, mysql_store_result should be default.